move permissions and dropdown definitions to Constants
