+++ imports/api/languages/service.js
@@ -1,7 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 
 import Users from '/imports/api/users/Users';
-import SystemLanguages from '/imports/api/dropdowns/SystemLanguages';
+import SystemLanguages from '/imports/framework/Constants/SystemLanguages';
 
 Meteor.methods({
   'language.get': () => {
