+++ imports/ui/dashboard/dashboard.details.js
@@ -1,6 +1,6 @@
 import { Template } from 'meteor/templating';
 
-import Permissions from '/imports/api/util/Permissions';
+import Permissions from '/imports/framework/Constants/Permissions';
 
 Template['dashboard.details'].helpers({
   data: {
